Moved HomeScreen to a separate folder
Reinstated the header.
