+++ src/components/screens/HomeScreen/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  barButtonItem: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    opacity: 0.75,
+    paddingHorizontal: 14,
+  },
+  barButtonImage: {
+    maxHeight: 44,
+    maxWidth: 44,
+  },
+});
